Simplified logic of disabling experience enhancer bar when a video ad is playing.

+++ victorious/victorious/UI/Content View/Experience Enhancers/VExperienceEnhancerController.h
@@ -14,7 +14,7 @@
 
 - (void)experienceEnhancersDidUpdate;
 
-@property (nonatomic, assign, readonly) float currentVideoTime;
+@property (nonatomic, assign, readonly) Float64 currentVideoTime;
 
 @end
